+++ api/src/Voucher.ts
@@ -1,7 +1,7 @@
+import { Option, Vec } from '@polkadot/types';
 import { BalanceOf } from '@polkadot/types/interfaces';
 import { HexString } from '@polkadot/util/types';
 import { ISubmittableResult } from '@polkadot/types/types';
-import { Option } from '@polkadot/types';
 import { SubmittableExtrinsic } from '@polkadot/api/types';
 import { blake2AsHex } from '@polkadot/util-crypto';
 
@@ -22,11 +22,18 @@ export class GearVoucher extends GearTransaction {
    *
    * @example
    * ```javascript
+   * import { VoucherIssued } from '@gear-js/api';
    * const programId = '0x..';
    * const account = '0x...';
-   * const { extrinsic, voucherId } = await api.voucher.issue(account, programId, 10000);
-   * extrinsic.signAndSend(account, (events) => {
-   *   events.forEach(({event}) => console.log(event.toHuman()));
+   * const { extrinsic } = await api.voucher.issue(account, programId, 10000);
+   * extrinsic.signAndSend(account, ({ events, status }) => {
+   *   if (status.isInBlock) {
+   *     const voucherIssuedEvent = events.find(({event: { method }}) => method === 'VoucherIssued')?.event as VoucherIssued;
+   *
+   *     if (voucherIssuedEvent) {
+   *       console.log('voucherId:', voucherIssuedEvent.data.voucherId);
+   *     }
+   *   }
    * })
    * ```
    */
@@ -84,7 +91,7 @@ export class GearVoucher extends GearTransaction {
    * const msgTx = api.message.send(...);
    * const tx = api.voucher.call(voucherId, { SendMessage: msgTx });
    * tx.signAndSend(account, (events) => {
-   *  events.forEach(({event}) => console.log(event.toHuman()));
+   *  events.forEach(({ event }) => console.log(event.toHuman()));
    * })
    * ```
    */
@@ -99,7 +106,8 @@ export class GearVoucher extends GearTransaction {
       return this._api.tx.gearVoucher.call(voucherId, {
         SendMessage: { destination, payload, gasLimit, value, keepAlive },
       });
-    } else if ('SendReply' in params) {
+    }
+    if ('SendReply' in params) {
       if (params.SendReply.method.method !== 'sendReply') {
         throw new Error(`Invalid method name. Expected 'SendReply' but actual is ${params.SendReply.method.method}`);
       }
@@ -107,7 +115,8 @@ export class GearVoucher extends GearTransaction {
       return this._api.tx.gearVoucher.call(voucherId, {
         SendReply: { replyToId, payload, gasLimit, value, keepAlive },
       });
-    } else if ('UploadCode' in params) {
+    }
+    if ('UploadCode' in params) {
       if (params.UploadCode.method.method !== 'uploadCode') {
         throw new Error(`Invalid method name. Expected 'UploadCode' but actual is ${params.UploadCode.method.method}`);
       }
@@ -117,6 +126,9 @@ export class GearVoucher extends GearTransaction {
       return this._api.tx.gearVoucher.call(voucherId, {
         UploadCode: { code },
       });
+    }
+    if ('DeclineVoucher' in params) {
+      return this._api.tx.gearVoucher.call(voucherId, { DeclineVoucher: null });
     }
 
     throw new Error('Invalid call params');
@@ -210,6 +222,15 @@ export class GearVoucher extends GearTransaction {
   }
 
   /**
+   * ### Decline existing and not expired voucher.
+   * @param voucherId The id of the voucher to be declined
+   * @returns Extrinsic to submit
+   */
+  decline(voucherId: string): SubmittableExtrinsic<'promise', ISubmittableResult> {
+    return this._api.tx.gearVoucher.decline(voucherId);
+  }
+
+  /**
    * ### Check if a voucher exists.
    * @param accountId
    * @param programId
@@ -256,33 +277,20 @@ export class GearVoucher extends GearTransaction {
   async getAllForAccount(accountId: string, programId?: HexString): Promise<Record<string, IVoucherDetails>> {
     const result: Record<string, IVoucherDetails> = {};
 
-    const keyPrefix = this._api.query.gearVoucher.vouchers.keyPrefix(accountId);
+    const voucherEntries = await this._api.query.gearVoucher.vouchers.entries(accountId);
 
-    const keysPaged = await this._api.rpc.state.getKeysPaged(keyPrefix, 1000, keyPrefix);
-
-    if (keysPaged.length === 0) {
+    if (voucherEntries.length === 0) {
       return result;
     }
 
-    const vouchers = (await this._api.rpc.state.queryStorageAt(keysPaged)) as Option<any>[];
+    for (const [key, value] of voucherEntries) {
+      const voucherId = key.args[1].toHex();
+      const details = value.unwrap();
 
-    vouchers.forEach((item, index) => {
-      const typedItem = this._api.createType<Option<PalletGearVoucherInternalVoucherInfo>>(
-        'Option<PalletGearVoucherInternalVoucherInfo>',
-        item,
-      );
+      const programs = details.programs.isSome ? (details.programs.unwrap().toJSON() as string[]) : null;
 
-      const voucherId = '0x' + keysPaged[index].toHex().slice(keyPrefix.length);
-
-      if (typedItem.isNone) {
-        return;
-      }
-
-      const details = typedItem.unwrap();
-      const programs = details.programs.unwrapOrDefault().toJSON() as string[];
-
-      if (details.programs.isSome && programId && !programs.includes(programId)) {
-        return;
+      if (programId && programs !== null && !programs.includes(programId)) {
+        continue;
       }
 
       result[voucherId] = {
@@ -291,7 +299,44 @@ export class GearVoucher extends GearTransaction {
         expiry: details.expiry.toNumber(),
         codeUploading: details.codeUploading.isTrue,
       };
-    });
+    }
+
+    return result;
+  }
+
+  /**
+   * ### Get all vouchers issued by an account.
+   * If many voucher are issued, this may take a while.
+   * @param accountId - The account id of the owner of the vouchers.
+   * @returns - An array of voucher ids.
+   */
+  async getAllIssuedByAccount(accountId: string): Promise<string[]> {
+    const result: string[] = [];
+
+    const keys = await this._api.query.gearVoucher.vouchers.keys();
+
+    if (keys.length === 0) {
+      return result;
+    }
+    for (let keyIndex = 0; keyIndex < keys.length; keyIndex += 1000) {
+      const vouchers = (await this._api.rpc.state.queryStorageAt(keys.slice(keyIndex, keyIndex + 1000))) as Vec<
+        Option<PalletGearVoucherInternalVoucherInfo>
+      >;
+
+      vouchers.forEach((info, index) => {
+        if (info.isNone) {
+          return;
+        }
+
+        if (!info.unwrap().owner.eq(accountId)) {
+          return;
+        }
+
+        const voucherId = keys[index + keyIndex].args[1].toHex();
+
+        result.push(voucherId);
+      });
+    }
 
     return result;
   }
@@ -312,7 +357,7 @@ export class GearVoucher extends GearTransaction {
 
     return {
       owner: owner.toHex(),
-      programs: programs.unwrapOrDefault().toJSON() as string[],
+      programs: programs.isSome ? (programs.unwrap().toJSON() as string[]) : null,
       expiry: expiry.toNumber(),
       codeUploading: codeUploading.isTrue,
     };