+++ api/test/Voucher.test.ts
@@ -111,7 +111,6 @@ describe('Voucher', () => {
     expect(txData.owner.toHuman()).toBe(alice.address);
 
     voucher = voucherId;
-    console.log(voucher, 'expires at', validUpTo);
   });
 
   test('Voucher exists', async () => {
@@ -127,6 +126,12 @@ describe('Voucher', () => {
     expect(vouchers[voucher]).toHaveProperty('codeUploading', true);
   });
 
+  test('Get all vouchers issued by account', async () => {
+    const vouchers = await api.voucher.getAllIssuedByAccount(alice.address);
+    expect(vouchers).toHaveLength(1);
+    expect(vouchers[0]).toBe(voucher);
+  });
+
   test('Get voucher details', async () => {
     const details = await api.voucher.getDetails(charlie.address, voucher);
     expect(details).toBeDefined();
@@ -135,8 +140,9 @@ describe('Voucher', () => {
     expect(details).toHaveProperty('expiry');
     expect(details).toHaveProperty('codeUploading');
     expect(Object.keys(details)).toHaveLength(4);
+    expect(details.programs).not.toBeNull();
     expect(details.programs).toHaveLength(1);
-    expect(details.programs[0]).toBe(programId);
+    expect(details.programs?.at(0)).toBe(programId);
     expect(details.owner).toBe(decodeAddress(alice.address));
     expect(details.expiry).toBe(validUpTo);
     expect(details.codeUploading).toBeTruthy();
@@ -231,16 +237,47 @@ describe('Voucher', () => {
     expect(txData.change.isActive).toBeTruthy();
   });
 
-  test.skip('Revoke voucher', async () => {
+  test('Decline voucher', async () => {
+    expect(voucher).toBeDefined();
+
+    const transferTx = api.balance.transfer(charlieRaw, 15 * 1e12);
+
+    await sendTransaction(transferTx, alice, ['Transfer']);
+
+    const tx = api.voucher.decline(voucher);
+
+    const [txData] = await sendTransaction(tx, charlie, ['VoucherDeclined']);
+
+    expect(txData).toBeDefined();
+    expect(txData).toHaveProperty('voucherId');
+    expect(txData).toHaveProperty('spender');
+    expect(Object.keys(txData.toJSON())).toHaveLength(2);
+    expect(txData.voucherId.toHex()).toBe(voucher);
+    expect(txData.spender.toHuman()).toBe(charlie.address);
+  });
+
+  test('Decline voucher with no funds', async () => {
+    const { extrinsic } = await api.voucher.issue(charlieRaw, 100e12, 1000, [programId], true);
+
+    const [txData] = await sendTransaction(extrinsic, alice, ['VoucherIssued']);
+
+    const voucherId = txData.voucherId.toHex();
+
+    const tx = api.voucher.call(voucherId, { DeclineVoucher: null });
+
+    const [txData2] = await sendTransaction(tx, charlie, ['VoucherDeclined']);
+
+    expect(txData2).toBeDefined();
+    expect(txData2).toHaveProperty('voucherId');
+    expect(txData2).toHaveProperty('spender');
+    expect(Object.keys(txData2.toJSON())).toHaveLength(2);
+    expect(txData2.voucherId.toHex()).toBe(voucherId);
+    expect(txData2.spender.toHuman()).toBe(charlie.address);
+  });
+
+  test('Revoke voucher', async () => {
     expect(voucher).toBeDefined();
     expect(validUpTo).toBeDefined();
-
-    let blockNumber = await api.rpc.chain.getHeader().then((header) => header.number.toNumber());
-
-    while (blockNumber < validUpTo) {
-      await sleep(3000);
-      blockNumber = await api.rpc.chain.getHeader().then((header) => header.number.toNumber());
-    }
 
     const tx = api.voucher.revoke(charlie.address, voucher);